flask_requirements.txt
python-dotenv==0.20.0
kafka-python==2.0.2
psycopg2-binary
Flask